+++ notify_send.py
@@ -49,7 +49,7 @@ SCRIPT_NAME = 'notify_send'
 SCRIPT_AUTHOR = 's3rvac'
 
 # Version of the script.
-SCRIPT_VERSION = '0.4 (dev)'
+SCRIPT_VERSION = '0.4'
 
 # License under which the script is distributed.
 SCRIPT_LICENSE = 'MIT'